third_party/blink/web_tests/platform/generic/fast/forms/select/optgroup-disabled-expected.txt

Test disabled attribute of optgroup element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check :disabled pseudo-class
PASS disabledSet["listbox_optgroup_disabled"] is true
PASS disabledSet["listbox_d1"] is true
PASS disabledSet["listbox_d2"] is true
PASS disabledSet["listbox_d3"] is true
PASS disabledSet["listbox_d4"] is true
PASS disabledSet["menulist_optgroup_disabled"] is true
PASS disabledSet["menulist_d1"] is true
PASS disabledSet["menulist_d2"] is true
PASS disabledSet["menulist_d3"] is true
PASS disabledSet["menulist_d4"] is true
PASS enabledSet["listbox_optgroup_enabled"] is true
PASS enabledSet["menulist_optgroup_enabled"] is true
Check IDL attribute
PASS $("listbox_optgroup_disabled").disabled is true
PASS $("listbox_optgroup_enabled").disabled is false
PASS $("menulist_optgroup_disabled").disabled is true
PASS $("menulist_optgroup_enabled").disabled is false
select.disabled doesn't affect optgroup.disabled
PASS $("listbox_optgroup_disabled").disabled is true
PASS $("listbox_optgroup_enabled").disabled is false
PASS $("menulist_optgroup_disabled").disabled is true
PASS $("menulist_optgroup_enabled").disabled is false
select.disabled doesn't affect pseudo-class :disabled
PASS disabledSet["listbox_optgroup_disabled"] is true
PASS disabledSet["menulist_optgroup_disabled"] is true
PASS disabledSet["listbox_optgroup_enabled"] is undefined.
PASS disabledSet["menulist_optgroup_enabled"] is undefined.
form.disabled doesn't affect optgroup.disabled
PASS $("listbox_optgroup_disabled").disabled is true
PASS $("listbox_optgroup_enabled").disabled is false
PASS $("menulist_optgroup_disabled").disabled is true
PASS $("menulist_optgroup_enabled").disabled is false
form.disabled doesn't affect pseudo-class :disabled
PASS disabledSet["listbox_optgroup_disabled"] is true
PASS disabledSet["menulist_optgroup_disabled"] is true
PASS disabledSet["listbox_optgroup_enabled"] is undefined.
PASS disabledSet["menulist_optgroup_enabled"] is undefined.
Check IDL [Reflect]
PASS $("listbox_optgroup_disabled").disabled is false
PASS $("menulist_optgroup_disabled").disabled is false
PASS $("listbox_optgroup_disabled").disabled is true
PASS $("menulist_optgroup_disabled").disabled is true
optgroup.disabled doesn't affect option.selected
PASS $("listbox").selectedIndex is 0
PASS $("listbox").selectedIndex is 4
PASS $("menulist").selectedIndex is 0
PASS $("menulist").selectedIndex is 4
PASS successfullyParsed is true

TEST COMPLETE

One
Two
Three
Four
One
Two
Three
Four


One
Two
Three
Four
One
Two
Three
Four
